+++ backend/ddang/src/main/java/com/ddang/ddang/report/presentation/AuctionReportController.java
@@ -6,8 +6,8 @@ import com.ddang.ddang.bid.presentation.resolver.LoginUser;
 import com.ddang.ddang.report.application.AuctionReportService;
 import com.ddang.ddang.report.application.dto.CreateAuctionReportDto;
 import com.ddang.ddang.report.application.dto.ReadAuctionReportDto;
-import com.ddang.ddang.report.presentation.dto.CreateAuctionReportRequest;
-import com.ddang.ddang.report.presentation.dto.ReadAuctionReportsResponse;
+import com.ddang.ddang.report.presentation.dto.request.CreateAuctionReportRequest;
+import com.ddang.ddang.report.presentation.dto.response.ReadAuctionReportsResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
